+++ lib/operators/misc/index.test.ts
@@ -2,12 +2,14 @@ import {
     $CovariancePop,
     $CovarianceSamp,
     $Expression,
+    $GetField,
     $Let,
     $Literal,
     $MergeObjects,
     $Meta,
     $Rand,
     $SampleRate,
+    $SetField,
 } from './';
 
 
@@ -32,8 +34,12 @@ describe('misc operators', () => {
         [$MergeObjects(...documents), { $mergeObjects: documents }],
         [$Meta(metaDataKeyword), { $meta: metaDataKeyword }],
         [$Let(vars, functionExpression), { $let: { vars, in: functionExpression } }],
-        [$CovariancePop(1, 2), { $covariancePop: [1, 2] }],
-        [$CovarianceSamp(2, 1), { $covarianceSamp: [2, 1] }],
+         [$CovariancePop(1, 2), { $covariancePop: [1, 2] }],
+         [$CovarianceSamp(2, 1), { $covarianceSamp: [2, 1] }],
+         [$GetField('fieldName'), { $getField: { field: 'fieldName' } }],
+         [$GetField('fieldName', { input: '$doc' }), { $getField: { field: 'fieldName', input: '$doc' } }],
+         [$SetField('fieldName', 'value'), { $setField: { field: 'fieldName', value: 'value' } }],
+         [$SetField('fieldName', 'value', { input: '$doc' }), { $setField: { field: 'fieldName', value: 'value', input: '$doc' } }],
     ])('should return %s', (
         operation: any,
         expected: any